Fix 17.x site build
The site plugin requires install to have run, lest it be unable to find
artifacts.  This makes sense since install is what actually puts them
into your repo.  The catch here is that install + site in the same build
hits issues with checkerqual, which conflicts with site.  The solution
is a two pass approach, which avoids the conflict.

We deliberately disabled the first build-only pass since we don't care
about the results of the test *and* we still run the tests in the site
pass.  If we don't build the tests during the site pass we don't get the
code coverage metrics that we're after with the site build in the first
place.