Allow `mvnw` to work with symlinks
We have `mvnw` to make running Maven easier. When compiling just one
module, it is somewhat annoying to type `../../mvnw` however. To solve
that, just create a symlink… except that doesn't work since the script
then thinks the symlink directory is the project's base directory.

That is fixed by this patch. It resolves symlink and will use the actual
path to the `mvnw` binary to determine the project's base path.