Remove or replace Function (and related classes)
Replaces or removes org.opencastproject.util.data.Function (and related classes).

Replaces with java.util.Function (or BiFunction or whatever makes sense).

Removes in cases where the method was clearly just wrapped in a Function
to be able to be used with Monadics. Streams don't require that, so those
methods can go back to being simple methods again.