Extended CA API - Add "stream" as a sample
Commit adds a new field "stream" next to "inputs". Shows all the files
that need to be changed in order to add a new field.

Includes changes necessary for the Admin UI to display more fields than
just inputs. Conversely, this will break the Admin UI without the
approppriate changes.