improve docs for execute-once WOH and execute service (#7108)
This PR improves the docs for the execute-once WOH. It now mentions the
existence of `#tags` for parameter substitution. In the execute service
docs it adds the missing `s` to the `tags` because the code actually
matches that. See
https://github.com/opencast/opencast/blob/develop/modules/execute-impl/src/main/java/org/opencastproject/execute/impl/ExecuteServiceImpl.java#L369

### How to test this patch

Use the execute-once WOH e.g. with `#{tags(engage-streaming)}` to select
the first track that matches.


### Your pull request should…

* [ ] have a concise title
* [ ] [close an accompanying
issue](https://docs.opencast.org/develop/developer/#participate/development-process/#automatically-closing-issues-when-a-pr-is-merged)
if one exists
* [ ] [be against the correct
branch](https://docs.opencast.org/develop/developer/development-process#acceptance-criteria-for-patches-in-different-versions)
* [ ] include migration scripts and documentation, if appropriate
* [ ] pass automated tests
* [ ] have a clean commit history
* [ ] [have proper commit messages (title and body) for all
commits](https://medium.com/@steveamaza/e028865e5791)
* [ ] explain why it needs to be merged into the legacy branch, if it is
targeting the legacy branch