Fix captions not published from editor
When adding captions to an event through the editor, and the event was not themed,
the captions would not be published. This fixes that.

Fixes the issue by adding another tag operation to the theming workflow that aims
to ensure that captions end up with the expected flavor (captions/shifted) after
the workflow has run.