Allow `secret` & `jwksUrl` to be unset leading to disabled login handler
Before, exactly one of those properties needed to be set. Now, both
being null is allowed, but the login handler disables itself then, not
accepting any tokens.

I see that this change might be a bit controversial or less nice, but
this enables us to uncomment most JWT related config in the security
config, leading to a drastically easier JWT setup.