Uncomment most JWT config in the security config for easier setup
In particular, this enables the header and query parameter filter by
default. This is safe and for most OC installations, both will be
required with upcoming changes to LMS plugins and Tobira.

But to enable both headers, the `jwtLoginHandler` also needs to exist
already, so it is also uncommented. This required the change in the last
commit, allowing a disabled `jwtLoginHandler` with neither `secret` nor
`jwksUrl` set. I also added a comment explaining that exactly one of
those have to be set to enable the login handler.

The JWKS config is moved down to the other cache/expiry related configs.