Add fallback fake email for JWT users
Not all JWTs contain a user email, in particular because it's often
not necessary. However, an email for a user is required apparently,
so I added this fake one as fallback. I first used
`['sub'] + '@jwt.invalid'` as fallback, but usernames might contain
'@' or other characters, making it an invalid mail.