Add new cutting "mode" to editor for vtt files
Currently, the editor removes cues from vtt files when the
cue overlaps with a cut segment. For example, if there is a
cue that goes from 10 to 16, and there is a cut from 13 to 20,
that cue would be removed. This patch adds an alternative "Shorten"
mode to the editor. With this active, cues are not removed but
their duration is reduced if possible. If the above example was
processed in shorten mode, the cue would remain and go from 10 to 13
instead. "Shorten" mode is intended to be used with chapters,
which have slightly different semantics from normal subtitles.